ref: 혼동을 피하기 위한 Consumer 일정 조회 메소드 이름 변경 및 DTO 위치 이동
