+++ homecare/src/main/java/jaega/homecare/domain/serviceMatch/dto/res/ConsumerNextScheduleResponse.java
@@ -1,15 +1,11 @@
-package jaega.homecare.domain.consumer.dto.res;
+package jaega.homecare.domain.serviceMatch.dto.res;
 
-import jaega.homecare.domain.serviceMatch.entity.MatchStatus;
 import jaega.homecare.domain.users.entity.ServiceType;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.UUID;
 
-public record ConsumerScheduleResponse(
-        UUID serviceRequestId,
-
+public record ConsumerNextScheduleResponse(
         String caregiverName,
 
         LocalDate serviceDate,
@@ -17,7 +13,6 @@ public record ConsumerScheduleResponse(
         LocalTime serviceEndTime,
 
         String serviceAddress,
-        ServiceType serviceType,
-        MatchStatus matchStatus
-
-){ }
+        ServiceType serviceType
+) {
+}
