+++ homecare/src/main/java/jaega/homecare/domain/serviceMatch/dto/res/ConsumerScheduleDetailResponse.java
@@ -1,0 +1,24 @@
+package jaega.homecare.domain.serviceMatch.dto.res;
+
+import jaega.homecare.domain.serviceMatch.entity.MatchStatus;
+import jaega.homecare.domain.users.entity.ServiceType;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+
+public record ConsumerScheduleDetailResponse(
+        String caregiverName,
+        String caregiverPhone,
+
+        LocalDate serviceDate,
+        LocalTime serviceStartTime,
+        LocalTime serviceEndTime,
+        Integer duration,
+
+        String serviceAddress,
+        ServiceType serviceType,
+        MatchStatus matchStatus,
+
+        Boolean hasReview               // 리뷰 존재 여부
+) {
+}
