+++ homecare/src/main/java/jaega/homecare/domain/serviceMatch/dto/res/ConsumerScheduleResponse.java
@@ -1,24 +1,23 @@
-package jaega.homecare.domain.consumer.dto.res;
+package jaega.homecare.domain.serviceMatch.dto.res;
 
 import jaega.homecare.domain.serviceMatch.entity.MatchStatus;
 import jaega.homecare.domain.users.entity.ServiceType;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.util.UUID;
 
-public record ConsumerScheduleDetailResponse(
+public record ConsumerScheduleResponse(
+        UUID serviceMatchId,
+
         String caregiverName,
-        String caregiverPhone,
 
         LocalDate serviceDate,
         LocalTime serviceStartTime,
         LocalTime serviceEndTime,
-        Integer duration,
 
         String serviceAddress,
         ServiceType serviceType,
-        MatchStatus matchStatus,
+        MatchStatus matchStatus
 
-        Boolean hasReview               // 리뷰 존재 여부
-) {
-}
+){ }
